ospfv3: implement authentication trailer feature
This commit introduces support for the authentication trailer feature
specified by RFC 7166.

For simplicity, the OSPF version-agnostic internal APIs use 64-bit
authentication sequence numbers, despite OSPFv2's use of 32-bit wide
authentication sequence numbers. Instead of creating a separate generic
type for authentication sequence numbers, truncating the u64 values
to u32 was deemed sufficient.

The implementation is currently limited to static keys, with support
for keychains planned for the future. All specified algorithms were
implemented: HMAC-SHA-1, HMAC-SHA-256, HMAC-SHA-384 and HMAC-SHA-512.

Interoperability tests were conducted successfully with the BIRD
routing stack. Unit tests were added for each HMAC-SHA variant.

It's worth noting that the implementation of this feature has led to
reduced coverage of ietf-ospf.yang due to the absence of authentication
trailer settings for virtual links, which are not yet supported.

+++ holo-ospf/src/error.rs
@@ -33,7 +33,7 @@ pub enum Error<V: Version> {
     InvalidDstAddr(V::NetIpAddr),
     PacketDecodeError(DecodeError),
     UnknownNeighbor(V::NetIpAddr, Ipv4Addr),
-    PacketAuthInvalidSeqno(V::NetIpAddr, u32),
+    PacketAuthInvalidSeqno(V::NetIpAddr, u64),
     InterfaceCfgError(String, V::NetIpAddr, PacketType, InterfaceCfgError),
     DbDescReject(Ipv4Addr, nsm::State),
     Ospfv2LsaUnknownType(V::LsaType),
